1.8 fix error

+++ include/gsbn/procedures/ProcFix/Pop.hpp
@@ -55,7 +55,7 @@ public:
 	SyncVector<fix16>* _bj;
 	SyncVector<fix16>* _dsup;
 	SyncVector<fix16>* _act;
-	SyncVector<int>* _spike;
+	SyncVector<int8_t>* _spike;
 	SyncVector<float>* _rnd_uniform01;
 	SyncVector<float>* _rnd_normal;
 	SyncVector<float>* _wmask;
